extracted ListHeader component
